backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
numpy==1.26.4
joblib==1.4.2
aiosmtplib==3.0.1
email-validator==2.1.1
